UIBase 슈퍼클래스
키보드 모드와 마우스 모드를 관리하는 공통 로직을 추출하여

슈퍼클래스인 UIBase에 넣음
